analyze_models REFACTOR: cache
